+++ defaults.py
@@ -123,7 +123,7 @@ class AppDefaults:
         "global_sel_line": '#0000ffbf',
         "global_alt_sel_fill": '#BBF268BF',
         "global_alt_sel_line": '#006E20BF',
-        "global_draw_color": '#FF000080',
+        "global_draw_color": '#00000080',
         "global_sel_draw_color": '#0000FF80',
 
         # Project Items colors