+++ CHANGELOG.md
@@ -7,6 +7,11 @@ CHANGELOG for FlatCAM Evo beta
 
 =================================================
 
+15.04.2022
+
+- in Geometry Editor moved the simplification feature in its own Editor Tool (plugin)
+- in Geometry Editor the default draw color is now black
+
 14.04.2022
 
 - in Geometry Editor added a new feature. While drawing a 'Path' now the user can project a direction by moving the mouse cursor in a certain direction and after that by typing a number or an arithmetic simple expression, a line segment will be drawn in that direction with the specified length from the last point